python_backend/requirements.txt

flask
openai
python-dotenv
